app is able to function
Deleted DUE DATE option. Will add that feature for the react version of the app. 
Users can add to-dos to their to-do list. If they completed a particular task, they can press 'complete' and the program will bring that particular task down to the 'completion list.' If the user presses the delete button either for the item on their to-do list or comp list, then the particular item will be deleted from that list.