combined retrieval of local storage functions
Combined the displayCompListOnRefresh() and displayTaskOnRefresh() into retrieveLocalStorageValue(keyName). The new function will be able to take the name of the key as an argument to check if there are any values in the local storage, and, if there is, it will get the to-do tasks and the completions and store them into the toDoArray and compList respectively.